templates/django-server/server/appname/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
djangorestframework-simplejwt>=5.3.0
redis>=5.0.1
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
django-redis>=5.4.0
celery
pipenv
